requirements.txt
selenium==3.141.0
requests==2.18.4
boto3==1.10.45
Flask==1.1.1
flask_wtf==0.14.2
greenstalk==1.0.1
Pillow==6.2.1
psycopg2==2.8.4
python-dotenv==0.10.3
WTForms==2.2.1
